pyproject.toml
[tool.poetry]
name = "botgpt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.17"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
slack-bolt = "^1.18.0"
openai = "^0.27.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
